Fix user_account, seller_account type as DECIMAL(10,2)

+++ database_setup.py
@@ -61,7 +61,7 @@ def create_tables():
             sex sex_type NOT NULL,
             email VARCHAR(100) NOT NULL UNIQUE,
             date_of_birth DATE,
-            user_account INT DEFAULT 10000
+            user_account DECIMAL(10, 2) DEFAULT 10000
         );
         """)
 
@@ -70,7 +70,7 @@ def create_tables():
             seller_id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
             seller_name VARCHAR(100) NOT NULL,
             contact_email VARCHAR(100) NOT NULL UNIQUE,
-            seller_account INT DEFAULT 0
+            seller_account DECIMAL(10, 2) DEFAULT 0
         );
         """)
